Changed Double To Float
refactor: 3d Shape was using doulbe instead of float for radius

+++ Project1/src/ThreeDimensionalShape.java
@@ -3,5 +3,5 @@ public abstract class ThreeDimensionalShape extends Shape {
         super(3);
     }
 
-    public abstract double calculateVolume();
+    public abstract float calculateVolume();
 }
